Separated countdown into a component to prevent Results from re-rendering.
