+++ front-end/src/components/Results.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useState } from "react";
+import NextGameCountdown from "./NextGameCountdown";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { Container, Row, Col } from "react-bootstrap";
 import { GiCheckMark } from "react-icons/gi"
@@ -13,8 +13,8 @@ function Results({ score, time, answers }) {
   for (let i = 0; i < answers.length; i++) {
     if (answers[i] === true) {
       timeBonus = Math.abs(time[i] - 15) * 10;
+      points += 100 + timeBonus;
     }
-    points += 100 + timeBonus;
   }
 
   // local storage
@@ -26,24 +26,6 @@ function Results({ score, time, answers }) {
   let totalGames = scoreArray.length;
   let highestScore = Math.max.apply(null, scoreArray);
   let averageScore = scoreArray.reduce((a, b) => a + b, 0) / scoreArray.length;
-
-  const [timeDifferenceHour, setTimeDifferenceHour] = useState("");
-  const [timeDifferenceMinute, setTimeDifferenceMinute] = useState("");
-  const [timeDifferenceSecond, setTimeDifferenceSecond] = useState("");
-  let nextDay = new Date();
-  nextDay.setHours(24, 0, 0, 0);
-  const midnight = nextDay.getTime() / 1000;
-  let currentTime = new Date().getTime() / 1000;
-
-  const countdown = setInterval( () => {
-      let timeDifferenceMS = midnight - currentTime;
-      setTimeDifferenceHour(Math.floor(timeDifferenceMS / 3600));
-      timeDifferenceMS = timeDifferenceMS - timeDifferenceHour * 3600;
-      setTimeDifferenceMinute(Math.floor(timeDifferenceMS / 60));
-      timeDifferenceMS = timeDifferenceMS - timeDifferenceMinute * 60;
-      setTimeDifferenceSecond(Math.floor(timeDifferenceMS));
-      clearInterval(countdown);
-  }, 1000)
 
   return (
     <>
@@ -78,13 +60,10 @@ function Results({ score, time, answers }) {
             </table>
             <h5>Your Score: {points} Points</h5>
             <h5>Total games played: {totalGames} </h5>
-            <h5>Highgest Score: {highestScore}</h5>
+            <h5>Highest Score: {highestScore}</h5>
             <h5>Average Score: {Math.round(averageScore)}</h5>
             <br></br>
-            <h5>Next SpeedTriv</h5>
-            <h6>Hours: {timeDifferenceHour}</h6>
-            <h6>Minutes: {timeDifferenceMinute}</h6>
-            <h6>Seconds: {timeDifferenceSecond}</h6>
+            <NextGameCountdown />
           </Col>
         </Row>
       </Container>
